SRP - moved computer move into separate class
