+++ lib/computer.rb
@@ -1,0 +1,5 @@
+class Computer  
+  def computer_move
+    weapon = ["rock", "paper", "scissors"].sample
+  end
+end
