+++ spec/features/game_spec.rb
@@ -1,7 +1,7 @@
 feature 'Entering a move' do
   scenario 'Player enters a move: rock, paper or scissors' do
     visit '/'
-    fill_in :name, with: 'Dhara'
+    fill_in :player_name, with: 'Dhara'
     click_button 'Submit'
     fill_in :player_move, with: 'rock'
     click_button 'Submit'
