+++ views/game.erb
@@ -1,4 +1,4 @@
-<h1> Welcome <%= @name %>! </h1>
+<h1> Welcome <%= @player_name %>! </h1>
 
 <h2> Please enter your move: rock, paper or scissors </h2>
 <form action="/game" method="post">
