+++ views/index.erb
@@ -1,6 +1,6 @@
 <h1>Please register your name below</h1>
 <form action="/names" method="post">
-  <label for="name">Name:</label>
-  <input type="text" name="name">
+  <label for="player_name">Name:</label>
+  <input type="text" name="player_name">
   <input type="submit" value="Submit">
 </form>
